requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==5.2
django-appconf==1.1.0
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environs==14.1.1
marshmallow==4.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.1
